+++ linepy/__init__.py
@@ -1,13 +1,13 @@
-from .client import LineClient
-from .channel import LineChannel
-from .poll import LinePoll
+from .client import LINE
+from .channel import Channel
+from .oepoll import OEPoll
 from akad.ttypes import OpType
 
-__copyright__       = 'Copyright 2017 by Fadhiil Rachman'
-__version__         = '2.0.2'
+__copyright__       = 'Copyright 2018 by Fadhiil Rachman'
+__version__         = '3.0.0'
 __license__         = 'BSD-3-Clause'
 __author__          = 'Fadhiil Rachman'
 __author_email__    = 'fadhiilrachman@gmail.com'
 __url__             = 'http://github.com/fadhiilrachman/line-py'
 
-__all__ = ['LineClient', 'LineChannel', 'LinePoll', 'OpType']
+__all__ = ['LINE', 'Channel', 'OEPoll', 'OpType']