+++ linepy/channel.py
@@ -1,69 +1,53 @@
 # -*- coding: utf-8 -*-
-from .client import LineClient
-from .timeline import LineTimeline
-from types import *
 
-class LineChannel(LineTimeline):
-    isLogin       = False
+def loggedIn(func):
+    def checkLogin(*args, **kwargs):
+        if args[0].isLogin:
+            return func(*args, **kwargs)
+        else:
+            args[0].callback.other("You want to call the function, you must login to LINE")
+    return checkLogin
+
+class Channel(object):
+    isLogin = False
     channelId     = None
-    profileDetail = None
+    channelResult = None
 
-    client      = None
-    server      = None
-    
-    channelAccessToken      = None
-    channelToken            = None
-    obsToken                = None
-    channelRefreshToken     = None
-    channelTokenExpiration  = None
-
-    def __init__(self, client, channelId=None):
-        if type(client) is not LineClient:
-            raise Exception("You need to set LineClient instance to initialize LineChannel")
+    def __init__(self, client, channelId, showSuccess=True):
         self.client = client
-        self.server = client.server
         self.channelId = channelId
-        self.login()
+        self.showSuccess = showSuccess
+        self.__loginChannel()
 
-    def login(self):
-        if self.channelId is None:
-            self.channelId=self.server.CHANNEL_ID['LINE_TIMELINE']
-        result = self.approveChannelAndIssueChannelToken(self.channelId)
-        
-        self.channelAccessToken     = result.channelAccessToken
-        self.channelToken           = result.token
-        self.obsToken               = result.obsToken
-        self.channelRefreshToken    = result.refreshToken
-        self.channelTokenExpiration = result.expiration
+    def __logChannel(self, text):
+        self.client.log('[%s] : Success login to %s' % (self.client.profile.displayName, text))
+
+    def __loginChannel(self):
         self.isLogin = True
+        self.channelResult  = self.approveChannelAndIssueChannelToken(self.channelId)
+        self.__createChannelSession()
 
-        self.createSession()
+    @loggedIn
+    def getChannelResult(self):
+        return self.channelResult
 
-    def createSession(self):
-        if self.isLogin:
-            self.server.setChannelHeadersWithDict({
-                'Content-Type': 'application/json',
-                'User-Agent': self.server.USER_AGENT,
-                'X-Line-Mid': self.client.profile.mid,
-                'X-Line-Carrier': self.server.CARRIER,
-                'X-Line-Application': self.server.APP_NAME,
-                'X-Line-ChannelToken': self.channelAccessToken
-            })
-            channelInfo = self.getChannelInfo(self.channelId)
-            if self.channelId == self.server.CHANNEL_ID['LINE_TIMELINE']:
-                LineTimeline.__init__(self)
-                self.profileDetail = self.getProfileDetail()
-                self.client.setChannelToModels(self)
-            self.client.log('[%s] : Success login to %s' % (self.client.profile.displayName, channelInfo.name))
+    def __createChannelSession(self):
+        channelInfo = self.getChannelInfo(self.channelId)
+        if self.showSuccess:
+            self.__logChannel(channelInfo.name)
 
+    @loggedIn
     def approveChannelAndIssueChannelToken(self, channelId):
-        return self.client.channel.approveChannelAndIssueChannelToken(channelId)
+        return self.client.approveChannelAndIssueChannelToken(channelId)
 
+    @loggedIn
     def issueChannelToken(self, channelId):
-        return self.client.channel.issueChannelToken(channelId)
+        return self.client.issueChannelToken(channelId)
 
+    @loggedIn
     def getChannelInfo(self, channelId, locale='EN'):
-        return self.client.channel.getChannelInfo(channelId, locale)
+        return self.client.getChannelInfo(channelId, locale)
 
+    @loggedIn
     def revokeChannel(self, channelId):
-        return self.client.channel.revokeChannel(channelId)
+        return self.client.revokeChannel(channelId)